Moved help toast to separate asset file
